+++ src/python/zensols/cli/app.py
@@ -612,10 +612,13 @@ class ApplicationFactory(PersistableContainer):
         actions: Tuple[ActionMetaData, ...] = tuple(chain.from_iterable(
             map(lambda a: a.meta_datas, cli_mng.actions.values())))
         config = CommandLineConfig(actions)
-        parser = CommandLineParser(config, self.package_resource.version,
-                                   default_action=cli_mng.default_action,
-                                   application_doc=self._get_app_doc(cli_mng),
-                                   usage_config=cli_mng.usage_config)
+        parser = CommandLineParser(
+            config,
+            self.package_resource.version,
+            default_action=cli_mng.default_action,
+            force_default=cli_mng.force_default,
+            application_doc=self._get_app_doc(cli_mng),
+            usage_config=cli_mng.usage_config)
         if logger.isEnabledFor(logging.DEBUG):
             logger.debug(f'created factory: {fac}')
         return fac, cli_mng, parser